requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
playwright>=1.40.0
requests>=2.31.0
